fix: widen loyalty point columns to numeric to match app's fractional multipliers
Tier multipliers (1.2x/1.5x) produce fractional loyalty points; int columns
rejected values like 4.5. users.loyalty_points and loyalty_transactions.points
are now numeric, matching what orderController and the seeder actually write.

+++ supabase/schema.sql
@@ -8,7 +8,7 @@ create table if not exists users (
   name text,
   phone text,
   birthday date,
-  loyalty_points int default 0,
+  loyalty_points numeric default 0,
   created_at timestamptz default now()
 );
 
@@ -53,7 +53,7 @@ create table if not exists loyalty_transactions (
   id text primary key,
   customer_id text,
   order_id text,
-  points int,
+  points numeric,
   transaction_type text,
   created_at timestamptz
 );
